harden defaultAuthStore a bit
Mostly this just prevents fs corruption, but it is a good practice
to at least set modes so it's not trivial for others to read the
token files.  Defense in depth.

Also use the auth type specified in token_type, with 'basic' and
'bearer' supported, just since that's what you're sposta do as a
well-behaved oauth device.  (Support for 'oauth' token type with roles
and grants and such can come later.)