build hybrid module for cjs/mjs ease
This makes it so that `import {Tier} from 'tier/client'` works even in
environments where there's no `exports` object.

The previous build was using TypeScript's CommonJS output mode, so it
assumed a Node target environment.  It would work if you had loaded
Deno's node polyfills, but that's not something we'd like to rely on.

PR-URL: https://github.com/tierrun/node-sdk/pull/3
Credit: @isaacs
Close: #3
Reviewed-by: @isaacs