feature.base can be fractional
tier.base still has to be an integer, even though Stripe has a field for
it, they don't support fractional prices for tier base prices.  Which
kind of makes sense, since a tier base price is always exactly 1 of that
base price, so it could never be an amount that's billable.